+++ jest/jest-mock.ts
@@ -1,13 +1,13 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 // @ts-nocheck
-import { AxiosRequestConfig } from 'axios';
+import { CreateAxiosDefaults } from 'axios';
 import { ApiKey } from '@getracker/api/types/lib/client';
 
 const api = {
     createClient: jest.fn(
-        (apiKey: ApiKey, opts?: AxiosRequestConfig | null) => true
+        (apiKey: ApiKey, opts?: CreateAxiosDefaults | null) => true
     ),
-    createClientFromEnv: jest.fn((opts?: AxiosRequestConfig | null) => true),
+    createClientFromEnv: jest.fn((opts?: CreateAxiosDefaults | null) => true),
 };
 
 module.exports = api;