+++ src/endpoints/index.ts
@@ -1,13 +1,10 @@
-import ActiveItemsWrapper from './ActiveItemsWrapper';
+import APIBaseWrapper from './APIBaseWrapper';
 import APIUptimeWrapper from './APIUptimeWrapper';
-import AuthWrapper from './AuthWrapper';
 import BarrowsRepairWrapper from './BarrowsRepairWrapper';
-import BillingWrapper from './BillingWrapper';
 import CombinationItemsWrapper from './CombinationItemsWrapper';
 import CraftingSmithingWrapper from './CraftingSmithingWrapper';
 import DashboardWrapper from './DashboardWrapper';
 import DecantPotionsWrapper from './DecantPotionsWrapper';
-import DeviceWrapper from './DeviceWrapper';
 import FavouriteItemsWrapper from './FavouriteItemsWrapper';
 import FletchingWrapper from './FletchingWrapper';
 import GELimitsWrapper from './GELimitsWrapper';
@@ -28,23 +25,18 @@ import ProfitTrackerWrapper from './ProfitTrackerWrapper';
 import RSUpdatesWrapper from './RSUpdatesWrapper';
 import StatsWrapper from './StatsWrapper';
 import StoreProfitWrapper from './StoreProfitWrapper';
-import SubscriptionsWrapper from './SubscriptionsWrapper';
 import SuggestedItemsWrapper from './SuggestedItemsWrapper';
 import TagWrapper from './TagWrapper';
-import TradingPostWrapper from './TradingPostWrapper';
 import UserWrapper from './UserWrapper';
 
 export {
-    ActiveItemsWrapper,
+    APIBaseWrapper,
     APIUptimeWrapper,
-    AuthWrapper,
     BarrowsRepairWrapper,
-    BillingWrapper,
     CombinationItemsWrapper,
     CraftingSmithingWrapper,
     DashboardWrapper,
     DecantPotionsWrapper,
-    DeviceWrapper,
     FavouriteItemsWrapper,
     FletchingWrapper,
     GELimitsWrapper,
@@ -65,9 +57,7 @@ export {
     RSUpdatesWrapper,
     StatsWrapper,
     StoreProfitWrapper,
-    SubscriptionsWrapper,
     SuggestedItemsWrapper,
     TagWrapper,
-    TradingPostWrapper,
     UserWrapper,
 };