+++ src/index.ts
@@ -5,21 +5,30 @@ import {
     DefaultOptions,
 } from './lib/client';
 import { getApiKeyFromEnv } from './lib/config';
-import { AxiosRequestConfig } from 'axios';
+import { CreateAxiosDefaults } from 'axios';
 
 const createInstance = (
     apiKey: ApiKey,
-    opts: AxiosRequestConfig | null = null
+    opts: CreateAxiosDefaults | null = null
 ): GeTrackerApi => {
     return createApi(createAxios(apiKey, opts));
 };
 
-const createClient = (apiKey: ApiKey, opts: AxiosRequestConfig | null = null) =>
-    createInstance(apiKey, opts);
-const createClientFromEnv = (opts: AxiosRequestConfig | null = null) =>
+const createClient = (
+    apiKey: ApiKey,
+    opts: CreateAxiosDefaults | null = null
+) => createInstance(apiKey, opts);
+const createClientFromEnv = (opts: CreateAxiosDefaults | null = null) =>
     createInstance(getApiKeyFromEnv(), opts);
 
 export * from './types';
 export * from './endpoints/types';
-export { createClient, createClientFromEnv, DefaultOptions };
+export { APIBaseWrapper } from './endpoints';
 export type { GeTrackerApi };
+export {
+    createClient,
+    createClientFromEnv,
+    createAxios,
+    DefaultOptions,
+    ApiKey,
+};